pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Django REST API backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.6",
    "django-filter>=24.3",
    "python-decouple>=3.8",
    "psycopg[binary]>=3.2",
    "gunicorn>=23.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10",
    "isort>=5.13",
    "flake8>=7.1",
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "factory-boy>=3.3",
    "pre-commit>=4.0",
    "django-debug-toolbar>=4.4",
]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip = [".venv", "migrations"]
known_django = ["django", "rest_framework"]
known_first_party = ["config", "core", "apps"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.coverage.run]
source = ["."]
omit = ["*/migrations/*", "*/tests/*", ".venv/*", "manage.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
